move charts to api
